+++ basic/tests/fixtures/data/config.php
@@ -16,6 +16,7 @@ return [
     ['key' => 'airline_name', 'value' => 'TestAirlines'],
     ['key' => 'no_reply_mail', 'value' => 'no-reply@testairlines.com'],
     ['key' => 'support_mail', 'value' => 'support@testairlines.com'],
+    ['key' => 'operations_mail', 'value' => 'operations@testairlines.com'],
 
     // Footer / Social
     ['key' => 'x_url', 'value' => 'https://testx.com/'],